Finally, the optimization algo is parallelized (for each batch).
The optimization is definitely unique enough that it doesn't
seem to be suitable to plug in a regular sgd implementation.

Reason for this is that, unlike word2vec where we are tuning
a single-layer neural net, we don't have a single set of params
to tune. Rather, each embedding becomes it's own set of params.