Fixed question mismatch between gini & best split.
- Resolved question mismatch between what was used to find the best gini and the
make_split function. The question was computed on the GPU for best_gini vs. on
the CPU for the make_split function. Due to FP mismatch, a slightly different value
was used which resulted in a split inconsistent with the gini hists iff there was a value
in the column sufficiently close to the question value. Soln: compute question on the GPU always.
- Removed loop for batch_evaluate_gini #batch_bins is currently nbins-1.