Implement (de)serialization of cuML's `Array`
Implements the `.serialize(...)` method for cuML `Array` objects. This
defers to cuDF's `Buffer.serialize(...)` for the initial serialization
pass. Then it fills in additional keyword arguments that the cuML
`Array` will need when being reconstructed. Finally it coerces all
frames to `Buffer`s explicitly.

As cuDF's `Buffer.deserialize(...)` is sufficiently general to work for
cuML's `Array`, this is simply relied upon in cuML.