Register `Array` with Dask for serialization
So that Dask knows cuML's `Array` can be serialized, register it as an
object than can be serialized using the functions provided in
`cuml.comm.serialize`. These just defer to the `.serialize(...)` and
`.deserialize(...)` methods of `Array`, which are no implemented. With
this registration, Dask will now dispatch through these methods anytime
a cuML `Array` is transmitted.