--singlegpu flag fix and calling script with only flags defined
Fixed issue with --singlegpu flag as it was not parsed correctly and as a result cuML didn't build properly for it.

Build script didn't run if only flags were specified. It should build the libraries and install them. Added check to see if complete build should be done (no arguments specified or then all arguments are flags).