Project Flash script changes (#2792)
This PR outlines the changes necessary for the new build process (aka Project Flash) to work with cuML. This PR is necessary to begin shadow testing of the new build process.

The goal is to simultaneously support both the old and new build processes. CI is configured to activate the new process by setting PROJECT_FLASH=1. Additionally BUILD_LIBCUML and BUILD_CUML will be set to 1 as needed by CI in the new process.

CI will stage build artifacts for python and GPU builds in ci/artifacts/cuml/cpu/ which the build scripts will then consume.