[REVIEW] Fix Stochastic Gradient Descent Example (#3136)
* Fix Stochastic Gradient Descent Example

The example that is currently in the docs does not run. dtype, penalty, lrate, loss are not defined. This new version sets the default values for the parameters of cumlSGD, and copies Mini Batch SGD Regression's dtype for pred_data['col1'], pred_data['col2']. When running this example, I also got slightly different values for the output, so these were also updated.

* Added PR #3136 to 0.17 Bug Fixes