Fix array_equal in tests (#3400)
Answers #1970
This PR fixes the `array_equal` function. More specifically, the calculation of the percentage of correct element in the array. This had as an effect a release in testing (necessary changes harder to detect). After verification, it appeared that the `total_tol` parameter was (explicitly or implicitly) set at default value everywhere. No change in pytests was necessary.

Authors:
  - Victor Lafargue (@viclafargue)

Approvers:
  - John Zedlewski (@JohnZed)

URL: https://github.com/rapidsai/cuml/pull/3400