Store data frequencies in tree nodes of RF (#3647)
The RF model should store the number of data points associated with each tree node. This information is useful in many ways, including:
* Visualizing the trees
* Debugging performance problem
* Computing SHAP values using the TreeSHAP algorithm

To that end, this PR does the following:
* Add the `instance_count` field in the `SparseTreeNode` structure
* Expose the `instance_count` field in the JSON dump
* Add a unit test to ensure that the counts in the JSON dump are correct.

Note that this feature will work with the new backend only. If the old backend is used, `instance_count` field will be absent in the JSON dump.

Closes #3131

Authors:
  - Philip Hyunsu Cho (@hcho3)

Approvers:
  - John Zedlewski (@JohnZed)

URL: https://github.com/rapidsai/cuml/pull/3647