Use clang-tools on x86 only (#3969)
Since there is a specific version of `clang-tools` we need and it may be difficult to migrate that to other architectures, restrict installing `clang-tools` to x86. After all this is only needed for checking code style, which can be done just on x86 without loss of generality.

Authors:
  - https://github.com/jakirkham

Approvers:
  - Jordan Jacobelli (https://github.com/Ethyling)
  - Corey J. Nolet (https://github.com/cjnolet)

URL: https://github.com/rapidsai/cuml/pull/3969