Raise ARIMA parameter limits from 4 to 8 (#4022)
Resolves #3915 

This PR still doesn't allow arbitrary parameters (because of the Jones transform kernel that uses the parameters as a template argument in order to work with registers).
But fixing the limit to 8 instead of 4 should be enough. In most cases, the parameters don't exceed 3 or 4, and more complex models don't give better results.

I also added tests with `p=5`.

Authors:
  - Louis Sugy (https://github.com/Nyrio)

Approvers:
  - Tamas Bela Feher (https://github.com/tfeher)
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/cuml/pull/4022