Improve numerical stability of the Kalman filter for ARIMA (#4259)
This PR improves the numerical stability of the Kalman filter by enforcing that the state covariance matrix is always symmetric and only has positive elements along its diagonal. Note that statsmodels does the former and not the latter, but I have found both to improve the stability in cases I was testing.

If stability problems arise in the future, a strategy to consider would be to use [Joseph form](https://en.wikipedia.org/wiki/Kalman_filter#Deriving_the_posteriori_estimate_covariance_matrix) for the covariance update. This avoids the subtraction which can result in a loss of positive-definiteness, but at the expense of performance because this formula is more complex.

Authors:
  - Louis Sugy (https://github.com/Nyrio)

Approvers:
  - Tamas Bela Feher (https://github.com/tfeher)
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/cuml/pull/4259