DBSCAN: Speed up adjgraph step
For large data sizes, the batch size of the DBSCAN algorithm is small in
order to fit the distance matrix in memory.

This results in a matrix that has dimensions num_points x batch_size,
both for the distance and adjacency matrix.

The conversion of the boolean adjacency matrix to CSR format is
performed in the 'adjgraph' step. This step was slow when the batch size
was small, as described in issue #2387.

In this commit, the adjgraph step is sped up. This is done in two ways:

1. The adjacency matrix is now stored in row-major batch_size x
   num_points format --- it was transposed before. This required changes
   in the vertexdeg step.

2. The csr_row_op kernel has been replaced by the adj_to_csr kernel.
   This kernel can divide the work over multiple blocks even when the
   number of rows (batch size) is small. It makes optimal use of memory
   bandwidth because rows of the matrix are laid out contiguously in
   memory.