Reduce Naive Bayes test time (#5082)
Linking #5053 
The number of columns is now reduced, a few similar tests are grouped together, to avoid fitting multiple time the same model on the same data. This results in time and memory usage reduction.

On my local machine the speedup is of x1.84, the duration goes from 142.79s (0:02:22) to 77.05s (0:01:17).

Authors:
  - Micka (https://github.com/lowener)
  - Dante Gama Dessavre (https://github.com/dantegd)

Approvers:
  - William Hicks (https://github.com/wphicks)

URL: https://github.com/rapidsai/cuml/pull/5082