Allowing large data in kmeans (#5228)
This is somewhat of an intermediate solution to fix an issue w/ kmeans that is currently prohibiting the algorithm from working when n_rows * n_cols > 2^32-1. 

For now, using 64-bit indexing allows the algorithm to work. We need to dig in further on the raft side to figure out where the numerical issue is happening. So far, some of this has to do w/ the new APIs we created and how it's handling indexing.

Authors:
  - Corey J. Nolet (https://github.com/cjnolet)
  - gpuCI (https://github.com/GPUtester)
  - https://github.com/Salonijain27

Approvers:
  - Divye Gala (https://github.com/divyegala)
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/cuml/pull/5228