Add rich HTML representation to estimators (#5630)
This adds a Jupyter (and other notebook) rich display hook that produces a HTML widget to represent an estimator in notebooks.

This adds the basics of having estimators displayed as HTML widgets in notebooks and other editors that use the Jupyter notebook "rich display" system.

<img width="304" alt="Screenshot 2023-10-26 at 15 33 37" src="https://github.com/rapidsai/cuml/assets/1448859/9eccb547-d37f-44b9-b284-fe284707764c">

This doesn't yet contain the cool feature of changing colour depending on fit status or the link to the documentation. For that we'd have to depend on a newer version of scikit-learn (or vendor the logic). In this case "newer" actually means "the next version to be released".

WDYT?

Authors:
  - Tim Head (https://github.com/betatim)
  - Simon Adorf (https://github.com/csadorf)
  - Dante Gama Dessavre (https://github.com/dantegd)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/5630