Update scikit-learn to 1.4 (#5851)
This is an attempt to update the scikit-learn dependency from 1.2 to 1.4. Most changes are related to constructor arguments that were deprecated in 1.2 and in 1.4 have changed/been removed.

A question I have is what cuml's deprecation policy is? I've gone with "two releases" for parameters where we can easily do so (deprecated in 24.06 and then remove them in 24.10). However that is only about 4 months of deprecation which could be a bit short.

Some of the changes would be hard to do as a deprecation (with 1.4 there is no way to provide the "old way"), we'd have to stick with 1.3 for now. I think this is a bit of a bummer but maybe the price to pay for not keeping on top of deprecations. And it seems like there is no deprecation policy in the docs/towards users? So maybe we can play this card once now, to catch up and at the same time introduce a deprecation policy.

The SHAP test needed its reference updating. I am not sure why, at least I couldn't quickly find a reason for why you'd have to do this.

I am not sure how possible it would be to support a range of scikit-learn versions (say 1.2 - 1.4). Would be cool but maybe not worth the added complexity?

Todo:
* [x] add deprecation warning in AgglomerativeClustering
* [ ] add tests for deprecations
	* [x] RF regressor
	* [x] RF classifier
	* [ ] ~~LARS~~ - LARS is experimental, so no need for deprecation
	* [x] LogisticRegression
	* [x] OneHotEncoder
	* [x] AgglomerativeClustering
* [ ] think about how to combine this with #5799
* [x] decide deprecation cycle length - copy cudf, so 24.06 -> 24.08
* [x] update "expiry" version in the warnings
* [x] update doc strings

xref #5799

Authors:
  - Tim Head (https://github.com/betatim)
  - Dante Gama Dessavre (https://github.com/dantegd)

Approvers:
  - Dante Gama Dessavre (https://github.com/dantegd)
  - Jake Awe (https://github.com/AyodeAwe)

URL: https://github.com/rapidsai/cuml/pull/5851