Make `get_param_names` a class method on single GPU estimators to match Scikit-learn closer (#6101)
Small difference between our estimators and Scikit-learn is that `get_param_names` are a classmethod in sklearn, and not in ours. This can make a few corner cases fail for using our estimators when Scikit-learn like estimators are expected. This PR fixes that. 

**Note:** This will not include dask-based estimators for the time being since they depend on introspection at object creation time.

Authors:
  - Dante Gama Dessavre (https://github.com/dantegd)
  - Divye Gala (https://github.com/divyegala)

Approvers:
  - William Hicks (https://github.com/wphicks)

URL: https://github.com/rapidsai/cuml/pull/6101