Remap BATCH_TREE_REORG to TREE_REORG in FIL
Avoid a bug which causes a batch of trees in the BATCH_TREE_REORG implementation to be labeled as complete (all having reached leaf nodes) when they are not. In the best case, this causes a crash when invalid shared memory is accessed. In the worst case, trees will silently contribute incorrect values to the final output.

This workaround simply disables BATCH_TREE_REORG by remapping BATCH_TREE_REORG to TREE_REORG in the C++ layer. The fundamental issue should be fixed through promotion of experimental FIL to default.