Skip dispatching to GPU for unimplemented metrics in UMAP (#6224)
This lists more metrics that the CPU umap library supports but cuml doesn't yet support. By listing them as not implemented we don't dispatch to the GPU when a user selects them.

Also added them to a test to check that they do not raise an error. It would be nice to check that they ran on the CPU when the accelerator is enabled, but I couldn't find a nice way to do it :-/ Ideas welcome.

Authors:
  - Tim Head (https://github.com/betatim)
  - Vyas Ramasubramani (https://github.com/vyasr)
  - Dante Gama Dessavre (https://github.com/dantegd)

Approvers:
  - James Lamb (https://github.com/jameslamb)
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/cuml/pull/6224