Plumb `metric` and `metric_kwds` through to UMAP `nn_descent`
Previously we were erroneously missing this plumbing, leaving the
`metric` and `metric_arg` as the defaults when `nn_descent` is used. In
the long run we should redo how the parameters are passed when
`nn_descent` is enabled to avoid this duplication, but for now fixing
the plumbing to be more correct seems fine.