Properly handle non-trivial classes in `LogisticRegression`
Scikit-learn's `LogisticRegression` contains support for non-trivial
classes (those that would typically require encoding before processing).

This PR adds support for that in both `fit` and `predict`. This is
complicated by `CumlArray`/`cupy`/`numba` not supporting non-numeric
types, which means we need to special case the output handling in
`predict`. It's gross, but functional.