Add new `cuml.accel` module (#6347)
This is just an alias shim around the old `cuml.experimental.accel` module, to help avoid merge conflicts or usage of the previous experimental module. In the future we should move the code out of experimental and over to `accel` instead.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - James Lamb (https://github.com/jameslamb)
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/cuml/pull/6347