Fix output type of `KernelRidge.predict`
Previously this wasn't handled correctly, but was _sometimes_ masked by
`GridSearchCV` in the old test fixing things behind the scenes.

This fixes that, adds a test that doesn't require the accelerator, and
changes the accelerator test to be both simpler (matching the smoketests
in the rest of the file) and much faster (the `GridSearchCV` in the test
made the test quite slow on my machine).