Remove `property` cases in `gpu_to_cpu`/`cpu_to_gpu`
These special cases should never be hit by working code, but caused
errors when properties weren't set yet. Checking if a class has an
attribute that's a property and manually calling the descriptor is
equivalent to `getattr(instance, attr)` but more convoluted.

This fixes things so that if a GPU class with a property matching a CPU
component isn't yet set, `gpu_to_cpu` doesn't error.