A few GPU<->CPU interop fixes (#6355)
- Fixes `gpu_to_cpu` and `cpu_to_gpu` to no longer error when moving fields from properties that don't _yet_ have a value. The previous special casing of `property` fields was unnecessary and should never have been able to be hit by working code (pulling the property from the class then calling `fget` manually is the same as a simple `getattr(instance, attr)`).
- Fixes `accelerator_active` to always be reset to its original value, even if `build_cpu_model` or `gpu_to_cpu` error. Previously if these methods errored the accelerator would be silently disabled for the rest of the run, leading to tricky to diagnose bugs in our test suite.

Both of these issues showed up after some changes made in #6346. I'm not really sure how to test the bugs in isolation (or if a test for that is even warranted). I can confirm that with rebasing #6346 on top of this PR the remaining issues in the test suite go away.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Dante Gama Dessavre (https://github.com/dantegd)

URL: https://github.com/rapidsai/cuml/pull/6355