Support non-trivial `classes_` in `LogisticRegression` (#6346)
Scikit-Learn's `LogisticRegression` is a bit unique in that it natively supports complex labels (e.g. raw strings/categories/non-monotonically increasing ints), rather than requiring that the labels are pre-encoded. It does this by internally using a `LabelEncoder` during `fit`, then converting the predicted labels back to the original label dtype in `predict`.

This PR adds support for this in `cuml`, improving compatibility with sklearn. This required:

- A small improvement to `LabelEncoder` to support all the input types that `cuml` natively supports.
- Addition of a `LabelEncoder` in `LogisticRegression.fit`
- Changing how we store `classes_`. Previously this used the descriptor functionality to support different container types. However, `CumlArray`/`cupy`/`numba` don't support non-numeric types, so we can't use that to store the classes anymore. We now always store `classes_` as a numpy array. I think this is fine - the size of classes is small - and also makes us a bit more compatible with sklearn since we can better ensure our dtypes match theirs.
- Changing `predict` to convert the numeric output back into the original classes. This was complicated since `CumlArray`/`cupy`/`numba` don't support non-numeric types, which means that most of our existing `output_type` machinery fails for these cases. I hacked something in that I think is sufficient, but it's definitely a hack.
- Addition of a new test to check things work properly across dtypes and container types.

This is an alternative to #6328. The fix here doesn't add any additional state, and I believe the test cases added here provide better coverage of the behavior we're trying to ensure.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)
  - Dante Gama Dessavre (https://github.com/dantegd)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6346