Ensure FIL CPU can be run without an available GPU
Due to an upstream change, CPU FIL was touching the CUDA context in a way that required a GPU to be present in order for it to be used. Until CPU FIL is actually included in the cuML CPU build (which will avoid this problem for anyone using the cuML CPU package), this change ensures that CPU FIL can still be run even if no GPU is available.