Centralize verbose <-> log level handling
Previously we had conversion to/from the `logger.level_enum` and
`verbose` parameters strewn across the codebase. We now centralize this
in 2 common utility functions. In the future we should also aim to
reduce the number of places where we need to handle this at all, but for
now this is a slight improvement.