Properly translate the `verbose` to `level_enum` in `simpl_set`
Previously we weren't translating `verbose` to `level_enum` in the
`simpl_set` functions. This meant that the default value (`False`)
would be cast to `level_enum`, resulting in the logger being set to `0`
(trace).

This doesn't fix the larger issue that our `verbose` handling mutates
the global logger settings. It does ensure we're at least translating
things correctly though.