Cleanup `cpu_to_gpu` and `gpu_to_cpu`
This is fixing a few warts I noticed when fixing a bug in some related
code a while ago. In particular, there's no reason to store the desired
order as an attribute on the cls as `{name}_order` - you can just pull
it off the descriptor directly when needed. The rest is some pythonic
improvements, but logically everything should be the same as it was
before.