Simplify `cpu_to_gpu` and `gpu_to_cpu` (#6395)
This is fixing a few warts I noticed when fixing a bug in some related code a while ago. In particular, there's no reason to store the desired order as an attribute on the cls as `{name}_order` (added [here](https://github.com/rapidsai/cuml/pull/4918/files#r1012248790)) - you can just pull it off the descriptor directly when needed. The rest is some pythonic improvements, but logic-wise everything should be the same as it was before.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Tim Head (https://github.com/betatim)
  - Victor Lafargue (https://github.com/viclafargue)

URL: https://github.com/rapidsai/cuml/pull/6395