Improve `accuracy_score` compatibility with `sklearn`.
This adds support for the `sample_weight` and `normalize` arguments to
`accuracy_score`. It also adds support for non-numeric dtypes (like
strings), which are support by sklearn in this API.

Like with the recent changes to  `r2_score`, this is done by moving the
implementation to use `cupy` instead of calling a C++ function.
`accuracy_score` is not a performance critical algorithm, and the `cupy`
implementation is both easier to maintain and should be good enough.

Also adds support for `sample_weight` in `ClassifierMixin.score`, fixing
a sklearn compatibility bug that was affecting the 0cc layer.