Plumb `metric` and `metric_kwds` through to UMAP with `nn_descent` (#6304)
Previously we were erroneously missing this plumbing, leaving the `metric` and `metric_arg` as the defaults when `nn_descent` is used. In the long run we should redo how the parameters are passed when `nn_descent` is enabled to avoid this duplication (there's a few other uselessly exposed params like `return_distances`), but for now fixing the plumbing to be more correct seems fine.

On top of #6303, leaving as draft for now until that's merged.

Also re-enables tests on ARM, fixes #5441.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Jinsol Park (https://github.com/jinsolp)
  - Victor Lafargue (https://github.com/viclafargue)
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6304