Add sklearn as run_constrained dep in cuml conda recipe
Currently `cuml` doesn't require scikit-learn is installed, but if it is
installed we recommend (and really require) it be a compatible version.

Short of making scikit-learn a hard dependency at this time, we instead
add `scikit-learn` as a `run_constrained` dependency with a lower bound
of 1.5.0 (released in May 2024). This is our recommended lowest
supported version of scikit-learn. If a user installs `cuml` with conda,
this will not necessarily pull in `scikit-learn` as well, but if they
also require `scikit-learn` this sets a constraint on the valid versions
the solver can install.