FIX Pass instances to the constructor of the meta-estimator in `SVC` (#6471)
The `estimator` attribute is used in the scikit-learn tags machinery to figure out what tags the meta estimator has. We pass a default constructed instance as it isn't actually used.

This was found as part of #6438 but can be fixed standalone. The problem isn't with the new scikit-learn version, but we discovered it with it.

cc @viclafargue

Authors:
  - Tim Head (https://github.com/betatim)

Approvers:
  - Jim Crist-Harif (https://github.com/jcrist)
  - Victor Lafargue (https://github.com/viclafargue)

URL: https://github.com/rapidsai/cuml/pull/6471