DOC Make `MultinomialNB` dask doctest more lax (#6482)
In https://github.com/rapidsai/cuml/actions/runs/13965838782/job/39097211126?pr=6438#step:10:2732 the test failed because the score was "a little to small". This makes the test a bit more forgiving in that it allows all scores that start with a `9`. We could rewrite the docstring to use something like `score() > 0.8` but I think the primary audience for a doctest is a human reader and for them the current version is probably more useful.

A somewhat unclear question for me is why the test failed in the first place, but 

xref #136

Authors:
  - Tim Head (https://github.com/betatim)

Approvers:
  - Victor Lafargue (https://github.com/viclafargue)
  - Jim Crist-Harif (https://github.com/jcrist)

URL: https://github.com/rapidsai/cuml/pull/6482