More permanently disable BATCH_TREE_REORG (#6374)
This PR is a follow-up to #6161 which more completely disables BATCH_TREE_REORG mode in legacy FIL. This requires more code change than expected because legacy FIL's public API is a fairly large surface and there are a number of ways that the `algo` parameter can be mutated or passed in at different locations. This should now completely disable any path that would lead to BATCH_TREE_REORG being selected and prevents compilation of the code path that actually triggers the bug in the BATCH_TREE_REORG implementation.

We do not entirely remove that code path from the source to avoid the need for more extensive changes to legacy FIL and its tests before we migrate to the new FIL code base.

Authors:
  - William Hicks (https://github.com/wphicks)
  - Philip Hyunsu Cho (https://github.com/hcho3)
  - Divye Gala (https://github.com/divyegala)

Approvers:
  - Divye Gala (https://github.com/divyegala)

URL: https://github.com/rapidsai/cuml/pull/6374