Remove `_experimental_dispatching` attribute (#6542)
This removes the unused `_experimental_dispatching` attribute. In the implementation this was never set to True, and exists from before `cuml.accel` was the way for enabling dispatching.

This also fully deletes a parametrized test testing our estimator dispatching. This was a convoluted parametrized test and would be better handled by the testing in `cuml/tests/accel`.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6542