Support `weights="distance"` for `KNeighbors*` in `cuml.accel`
Previously we would fail if the user specified `weights="distance"` to
`KNeighborsClassifier`/`KNeighborsRegressor`. This fixes that and adds a
test.

Part of fixing this required changing the logic in `dispatch_func` to
not special-case inference methods. Previously we would always run
inference on the GPU, even if `_gpuaccel` was False (meaning the
hyperparameters specified weren't supported by cuml). I don't believe
this to be the desired logic - if cuml doesn't support the specified
hyperparameters, we cannot be sure that we do support them for
`predict`. Further, the state is stored on the cpu estimator already,
running inference on CPU makes more sense anyway IMO. It also makes
understanding where something runs clearer:

- If the hyperparameters aren't supported by cuml, then everything runs
  on CPU
- If the arguments provided to a method aren't supported by cuml, then
  that method will dispatch to CPU
- Otherwise we run on GPU