Don't exclude sklearn tests from acceleration
When we rewrote `cuml.accel`'s mechanism to use a finder and loader (and
include a `denylist` for excluding certain modules from acceleration),
we accidentally excluded the `sklearn` tests from acceleration. We don't
want to do this since we run the sklearn test suite with `cuml.accel`
enabled to validate our implementation.

This PR:
- Renames `denylist` to `exclude`
- Makes `exclude` optionally take a callable instead of a sequence
- Updates our exclude logic for `cuml.accel` to exempt modules matching
  `sklearn.*.tests.*` from exclusion.