Update SVC tests for CCCL update (#6569)
This test was failing after the introduction of CCCL 2.8 in https://github.com/rapidsai/cuml/pull/6421. The cause was [a non-deterministic minimum search in the SmoBlockSolve kernel](https://github.com/rapidsai/cuml/blob/b01e2d1d4cf358dbc0a435a663c546fa9cf3e3cb/cpp/src/svm/smoblocksolve.cuh#L206) with   the `cub::BlockReduce` feature of the CCCL library. During the first iteration of the first launch of the kernel, the values being searched are set to -1.0 making the result of the search undefined/non-deterministic. The minimum search in this kernel produces a different result on this first iteration when libcuml is compiled with CCCL 2.8 rather than version 2.7 (probably because of underlying algorithms updates) making the training of SVC produce different results..
 
Since the value of the `gamma` hyper-parameter does not seem to be critical to this pytest test, I modified its value to stabilize the test. Also updated the expected values in the SVC gtests.

Authors:
  - Victor Lafargue (https://github.com/viclafargue)
  - Divye Gala (https://github.com/divyegala)

Approvers:
  - Divye Gala (https://github.com/divyegala)

URL: https://github.com/rapidsai/cuml/pull/6569