Remove guarded imports of `scipy`
SciPy is currently a required dependency of cuml (and has been for a
while). With the intent to make sklearn a required dependency (which
also requires scipy), it seems unlikely we'll ever step back from this
requirement.

This PR removes all the remaining guarded imports of scipy, along with
the `has_scipy` function. There were plenty of places where we didn't
guard the import, so doing this improves consistency and deletes some
unneeded branching. In the 2 places where we were using `has_scipy` to
also check the version I've inlined the checks which is also more
readable IMO.