FIX Propagate random state to numpy rng in `make_classification` (#6518)
When generating a classification problem a mix of cupy and numpy random generators is used. The random state needs to be propagated to the numpy generator which was not happening.

I'm not sure what the best way of doing the "propagation" is. There is prior art for using `randint` and then using that as a seed. We don't need the sequences to be the same, we just need a way to initialise the Numpy random generator. 

An alternative would be to use the cupy rng throughout and combine and it with moving the data to a numpy array.

closes #6510

Authors:
  - Tim Head (https://github.com/betatim)
  - Simon Adorf (https://github.com/csadorf)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6518