Do not set `n_features_in_` attribute before `fit` (#6624)
Fixes #6528

Scikit-learn compatible estimators should not set fitted attributes (those ending in a `_`) before they are fit. In the reported case the problem occurs because `skore` uses `check_is_fitted` from scikit-learn to decide if the estimator passed in is already fitted or not. This check looks for the presence of fitted attributes. This means the random forest model appeared to be fitted, when it wasn't.

What is the right label for a bug fix like this? While it does "break" existing user code that erroneously relies on this attribute always existing, that is buggy code as well. Hence I'd classify it as a bug fix. But it seems like I can only choose between non-breaking and breaking?

Authors:
  - Tim Head (https://github.com/betatim)
  - Simon Adorf (https://github.com/csadorf)

Approvers:
  - Victor Lafargue (https://github.com/viclafargue)
  - Jim Crist-Harif (https://github.com/jcrist)

URL: https://github.com/rapidsai/cuml/pull/6624