Remove cuml-cpu and CPU <> GPU docs
Previously cuml had experimental support for CPU <> GPU operation. This
came in two forms:

- `cuml-cpu`: a CPU-only build of cuml
- Setting the device type to CPU using `using_device_type` or
  `set_global_device_type`

The former was deprecated and has been fully removed from the codebase.

The latter is being ripped out (without deprecation but with warnings
that setting the device type no longer does anything) as we redo the
cuml-accel layer. Execution on CPU in a GPU-enabled build was fairly
buggy, only certain methods properly worked, and touching the wrong
thing would lead to immediate segfaults. Considering we haven't gotten
any bug reports for this, my guess is this feature was mostly unused
anyway.

This PR removes the documentation for cuml-cpu (already removed) and
`using_device_type` (discouraged, will be going away in the future).

Future users should either:
- Use `cuml.accel` to run the same code on GPU or CPU
- Use `as_sklearn`/`from_sklearn` to convert cuml estimators to/from
  their CPU counterparts. This allows workflows like fitting on GPU and
  inference on CPU, and is the recommended pattern going forward.