Update xfail list
Sklearn generates the parameterized tests in `test_common` dynamically.
Before when our proxy estimators weren't subclasses of `BaseEstimator`
they were fully ignored by these tests. Now that they're true
`BaseEstimator` subclasses, there are suddenly a lot more tests for
them. We pass _most_ of the checks, the ones that we don't have been
added to the xfail list.

There are a few other newly failing tests. I went through them
individually - in most cases the issue was that before we had a bug, and
that bug just happened to make the test not fail. A good example of this
is the `ElasticNet` tests - there's a few that run `ElasticNet` with two
different sets of hyperparameters and compare the results. Before we
were silently dropping the relevant hyperparameters, so both estimators
were identical and would of course pass. Now they're not, and we don't
have identical `coef_` since one is running on GPU and one on CPU. But
the scores of both estimators are comparable.

We also remove several xfailed tests. All in all, I think this PR is a
net positive, and afaict doesn't introduce new bugs per the sklearn test
suite.