Expand sklearn matrix used in tests (#6698)
Previously we were only running one run of the sklearn test suite in PRs. Now that we support multiple versions of sklearn, we amend this list to select one run with the oldest deps and one run with the latest deps.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Tim Head (https://github.com/betatim)
  - Gil Forsyth (https://github.com/gforsyth)

URL: https://github.com/rapidsai/cuml/pull/6698