FIX Allow prereleases when evaluating xfail conditions (#6722)
To test `cuml.accel` with the scikit-learn test suite we maintain a list of expected failures which can contain conditionals.

This PR makes it so that a condition like `scikit-learn >= 1.6` will evaluate to true when scikit-learn 1.7.0rc0 is installed. This is useful when we want to test a upcoming release of scikit-learn (or other packages) for `cuml.accel`.

Authors:
  - Tim Head (https://github.com/betatim)
  - Simon Adorf (https://github.com/csadorf)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6722