Move FIL docs to user docs (#6762)
Closes #6492 

Moves the FIL docs to the correct place in the cuML user documentation. 

PR converts the markdown file to RST, moves it and adds it to the index, but does not modify the actual contents of the documentation. We can do any changes in further PRs as other FIL-related PRs are merged in.

Authors:
  - Dante Gama Dessavre (https://github.com/dantegd)

Approvers:
  - Simon Adorf (https://github.com/csadorf)

URL: https://github.com/rapidsai/cuml/pull/6762