Mark `test_equivalence_components_pca_spca` to flaky
This was previously marked as failing - I've now seen it pass a few
times. Marking as flaky instead of passing for now while the rest of
cuda 12.9 shakes out. As noted in the comment, it's not _wrong_ that the
test fails, it's just the signs being inverted on some of the vectors.