Mark `test_equivalence_components_pca_spca` as flaky (#6834)
This was previously marked as failing - I've now seen it pass a few times. Marking as flaky instead of passing for now while the rest of cuda 12.9 shakes out. As noted in the comment, it's not _wrong_ that the test fails, it's just the signs being inverted on some of the vectors.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Simon Adorf (https://github.com/csadorf)
  - Tim Head (https://github.com/betatim)

URL: https://github.com/rapidsai/cuml/pull/6834