Fix several bugs in `cuml.accel` CLI (#6832)
This is a general cleanup of the `cuml.accel` CLI, fixing several bugs and adding tests:

- Fixes a bug where `-m module` didn't work at all
- Fixes a bug where trailing arguments wouldn't properly be forwarded to modules or scripts if they conflicted with options/arguments to `cuml.accel` itself.
- Fixes a bug where `--verbose` would be overwritten by the later `install` call, meaning the log level was always stuck at `INFO` (the default for `install`). I've changed it to default to `WARN` in `install`, and ensured the log-level setting logic always delegates to that function rather than having a separate call in the CLI.
- Fixes a bug where executing from `stdin` would silently ignore errors, leading failing scripts to appear as if they passed successfully.
- Fixes a bug where `--convert-to-sklearn` didn't do the right thing anymore under the new `ProxyBase` estimators.

I've added tests for all this to ensure we have full coverage. I've also taken the time to improve the CLI help documentation so it's hopefully more approachable for users.

This was all motivated by trying to ensure that stacking accelerators worked in the CLI as well as it does with ipython magics.

```
python -m cuml.accel -m cudf.pandas ...
```

This led to finding and fixing a whole slew of bugs. Once this is in I have few additional followups I'd like to make:

- Deprecate the `--cudf-pandas` flag. Stacking accelerators is more composable and now works fine.
- Deprecate the `--convert-to-sklearn` flags. The new `ProxyBase` classes eliminate the need for it entirely.
- Add support for the `-c` flag as well, to mirror the standard python interpreter and `cudf.pandas` implementation.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Tim Head (https://github.com/betatim)

URL: https://github.com/rapidsai/cuml/pull/6832