Port UMAP to InteropMixin/ProxyBase
This ports `cuml.UMAP` to subclass from `InteropMixin` instead of
`UniversalBase`, and switches to using `ProxyBase` to define the
respective proxy estimator.

This turned up a few gross bugs and edge cases with the previous
implementation. I've opted to d testing as I think makes sense here
covering the most common operating paths. I believe this to be strictly
better than what we had before, but don't doubt there's still cases we
don't handle smoothly. For best results we'll want to start running the
upstream umap-learn test suite, but IMO this falls outside the scope of
this PR.