Remove `GlobalSettings.device_type`
This removes the global `device_type` setting and all downstream
consumers of it. As of 25.06 `GlobalSettings.device_type` is always
`DeviceType.device`, and all methods of changing it have been deprecated
and removed. This API is no longer needed, removing it is another step
towards stripping out the remnants of `cuml-cpu`.