Rewrite `HDBSCAN` python wrapper (#6913)
This is pretty much a full rewrite of the python side of our `HDBSCAN` wrapper. There were a few goals here:

- Fix #6541
- Significantly simplify and reduce the state stored on an instance of `HDBSCAN` to ease #6711
- Isolate the different code paths for initializing state (e.g. from fit, from sklearn, unpickling, ...) to better ensure we always have a valid `HDBSCAN` instance (and that resources are being properly tracked and freed)
- Remove several cases of poor cython practice (e.g. allocating something via `new` then passing it around via a python `int` before later freeing)

The changes I've made here accomplish all these goals. There are still further simplifications that could be made, but those will require some API changes to the C++ layer. For now I've punted on those, once this is in I'll open some follow-up issues to discuss further improvements.

Note that for the `cuml.accel` layer we're not in a fully correct state _yet_. We're better off than we were before (and all tests continue to pass), but for `sklearn <> cuml` interop to be fully correct we'll want to finish up #6711. I've left that work out of this PR to minimize the already large diff.

Fixes #6541.

Authors:
  - Jim Crist-Harif (https://github.com/jcrist)

Approvers:
  - Divye Gala (https://github.com/divyegala)

URL: https://github.com/rapidsai/cuml/pull/6913